Finished Milestone 1. Added documentation for the program.
